make plotting functionality an extension package add docs deps
