+++ docs/src/index.md
@@ -1,47 +1,55 @@
 ```@raw html
 <div align="center">
-
-<a href="https://github.com/kestrelquantum/Piccolo.jl">
-  <img src="assets/logo.svg" alt="logo" width="25%"/>
-</a> 
-
-<div style="display: table; width: 100%;">
-  <div style="display: table-row;">
-    <div style="display: table-cell; text-align: center;"><b>Documentation</b></div>
-    <div style="display: table-cell; text-align: center;"><b>Build Status</b></div>
-    <div style="display: table-cell; text-align: center;"><b>Support</b></div>
-  </div>
-  <div style="display: table-row;">
-    <div style="display: table-cell; text-align: center;">
-      <a href="https://kestrelquantum.github.io/NamedTrajectories.jl/stable/">
-        <img src="https://img.shields.io/badge/docs-stable-blue.svg" alt="Stable"/>
-      </a>
-      <a href="https://kestrelquantum.github.io/NamedTrajectories.jl/dev/">
-        <img src="https://img.shields.io/badge/docs-dev-blue.svg" alt="Dev"/>
-      </a>
-    </div>
-    <div style="display: table-cell; text-align: center;">
-      <a href="https://github.com/kestrelquantum/NamedTrajectories.jl/actions/workflows/CI.yml?query=branch%3Amain">
-        <img src="https://github.com/kestrelquantum/NamedTrajectories.jl/actions/workflows/CI.yml/badge.svg?branch=main" alt="Build Status"/>
-      </a>
-      <a href="https://codecov.io/gh/kestrelquantum/NamedTrajectories.jl">
-        <img src="https://codecov.io/gh/kestrelquantum/NamedTrajectories.jl/branch/main/graph/badge.svg" alt="Coverage"/>
-      </a>
-    </div>
-    <div style="display: table-cell; text-align: center;">
-      <a href="https://unitary.fund">
-        <img src="https://img.shields.io/badge/Supported%20By-Unitary%20Fund-FFFF00.svg" alt="Unitary Fund"/>
-      </a>
-    </div>
-  </div>
+  <a href="https://github.com/kestrelquantum/Piccolo.jl">
+    <img src="assets/logo.svg" alt="logo" width="25%"/>
+  </a> 
 </div>
 
-<br>
-<i>An elegant way to handle messy trajectory data</i>
-<br>
-
+<div align="center">
+  <table>
+    <tr>
+      <td align="center">
+        <b>Documentation</b>
+        <br>
+        <a href="https://kestrelquantum.github.io/NamedTrajectories.jl/stable/">
+          <img src="https://img.shields.io/badge/docs-stable-blue.svg" alt="Stable"/>
+        </a>
+        <a href="https://kestrelquantum.github.io/NamedTrajectories.jl/dev/">
+          <img src="https://img.shields.io/badge/docs-dev-blue.svg" alt="Dev"/>
+        </a>
+      </td>
+      <td align="center">
+        <b>Build Status</b>
+        <br>
+        <a href="https://github.com/kestrelquantum/NamedTrajectories.jl/actions/workflows/CI.yml?query=branch%3Amain">
+          <img src="https://github.com/kestrelquantum/NamedTrajectories.jl/actions/workflows/CI.yml/badge.svg?branch=main" alt="Build Status"/>
+        </a>
+        <a href="https://codecov.io/gh/kestrelquantum/NamedTrajectories.jl">
+          <img src="https://codecov.io/gh/kestrelquantum/NamedTrajectories.jl/branch/main/graph/badge.svg" alt="Coverage"/>
+        </a>
+      </td>
+      <td align="center">
+        <b>License</b>
+        <br>
+        <a href="https://opensource.org/licenses/MIT">
+          <img src="https://img.shields.io/badge/License-MIT-yellow.svg" alt="MIT License"/>
+        </a>
+      </td>
+      <td align="center">
+        <b>Support</b>
+        <br>
+        <a href="https://unitary.fund">
+          <img src="https://img.shields.io/badge/Supported%20By-Unitary%20Fund-FFFF00.svg" alt="Unitary Fund"/>
+        </a>
+      </td>
+    </tr>
+  </table>
 </div>
 
+<div align="center">
+  <i>An elegant way to handle messy trajectory data</i>
+  <br>
+</div>
 ```
 
 # NamedTrajectories.jl
