+++ test/test_utils.jl
@@ -22,18 +22,6 @@ function get_free_time_traj(;
     return NamedTrajectory(free_time_data; timestep=Δt, controls=:u, kwargs...)
 end
 
-function get_fixed_time_traj2(;
-    T::Int=5, 
-    Δt::Float64=0.1, 
-    x_dim::Int=3, 
-    y_dim=3, 
-    a_dim::Int=2,
-    kwargs...
-)
-    fixed_time_data = (x = rand(x_dim, T), y_dim = rand(y_dim, T), u = rand(a_dim, T))
-    return NamedTrajectory(fixed_time_data; timestep=Δt, controls=:u, kwargs...)
-end
-
 function named_trajectory_type_1(; free_time=false)
     # Hadamard gate, two dda controls (random), Δt = 0.2
     data = [
